refactor: routes rename
