+++ src/app/lib/constants.js
@@ -1,0 +1,5 @@
+const ICON_URL = 'https://raw.githubusercontent.com/ringcentral/trello-notification-app/main/icons/trello.png';
+const COLOR_IMAGE_URL_BASE = 'https://raw.githubusercontent.com/ringcentral/trello-notification-app/main/icons/';
+
+exports.ICON_URL = ICON_URL;
+exports.COLOR_IMAGE_URL_BASE = COLOR_IMAGE_URL_BASE;
